docs: document vfx engines
